refactor: 리팩토링 및 컴포넌트 구조 개선, ScheduleTable 및 SearchDialog 통합
