+++ src/components/SearchDialog/filters/DayFilter.tsx
@@ -1,6 +1,6 @@
 import { Checkbox, CheckboxGroup, FormControl, FormLabel, HStack } from "@chakra-ui/react";
 import { memo } from "react";
-import { DAY_LABELS } from "../../../constants";
+import { DAY_LABELS } from "../../../constants/constants";
 
 interface DayFilterProps {
   days: string[];
