+++ src/components/SearchDialog/filters/TimeFilter.tsx
@@ -11,7 +11,7 @@ import {
   Wrap,
 } from "@chakra-ui/react";
 import { memo } from "react";
-import { SearchOption } from "../../../types";
+import { SearchOption } from "../../../types/types";
 
 const TIME_SLOTS = [
   { id: 1, label: "09:00~09:30" },
