+++ src/components/Table/ScheduleTables.tsx
@@ -1,10 +1,10 @@
 import { Button, ButtonGroup, Flex, Heading, Stack } from "@chakra-ui/react";
 import { useState } from "react";
-import { useScheduleContext } from "./ScheduleContext.tsx";
+import { useScheduleContext } from "../../provider/ScheduleContext.tsx";
 import ScheduleTable from "./ScheduleTable.tsx";
-import SearchDialog from "./SearchDialog.tsx";
-import ScheduleDndProvider from "./ScheduleDndProvider.tsx";
-import { useAutoCallback } from "./hooks/useAutoCallback.ts";
+import SearchDialog from "../SearchDialog/SearchDialog.tsx";
+import ScheduleDndProvider from "../../provider/ScheduleDndProvider.tsx";
+import { useAutoCallback } from "../../hooks/useAutoCallback.ts";
 
 export const ScheduleTables = () => {
   const { schedulesMap, setSchedulesMap } = useScheduleContext();
